tests/wpt/web-platform-tests/css/css-contain/content-visibility/content-visibility-080.html

<!doctype HTML>
<html id=html>
<meta charset="utf8">
<title>Content Visibility: caret position with html hidden</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<meta name="assert" content="caretRangeFromPoint works even if html has content-visibility hidden">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<meter></meter>
<iframe></iframe>
<style>
* {
  all: initial;
  content-visibility: hidden;
}
</style>

<script>
test((t) => {
  t.add_cleanup(() => { document.getElementsByTagName('style')[0].remove(); });
  const range = document.caretRangeFromPoint();
  assert_not_equals(range, null, "range exists");
  assert_equals(range.startContainer, html, "startContainer is html");
  assert_equals(range.startOffset, 0, "startOffset is zero");
  assert_equals(range.endContainer, html, "endContainer is html");
  assert_equals(range.endOffset, 0, "endOffset is zero");
}, "Caret range from point");
</script>
</html>
